Plan out UserReceiver replacement
This is going to be the new method of handling the game monitor. Instead of relying on guild-side receivers, it will instead be bound to the user's direct messaging system. This will be much easier to update displays across everyone, as well as being able to manage possible rate limitations that could occur when too many users send an update packet to a receiver at once.