Refactor context display bodies
The way display information was received and constructed was sloppy and missed out on extra information that might be needed. This new version now allows you to see the family tree of context values and itself as a generic object. The report system was also cleaned up to allow for a generic variant.