Start game framework tweaks and user services
This was a minor step in creating the frameworks for handling users. This will now manage merits, balances, stats, upgrades, attributes, and cooldowns. On a side note, the game framework was tweaked yet again to properly assign builder objects into a game builder. In this case now, custom games can be easily implemented.