Help service rework: 0.0.9-dev
This major update focuses on completely reworking the help service to be even easier to use. All of the services, display information models, and other utility classes have been updated to reflect that. Even more so, other areas of the bot were cleaned up a bit to provide easier code readability.