Create InfoService final: v0.0.12
The InfoService was completely rewritten to finalize and add support for guides and reports. Tooltips are now configurable, and new attributes were created to simply command structures. The user and guild account systems were entirely rewritten to their final structure, as part of the preparation for v1.0.0. The saving mechanic used for accounts was also tweaked, so that initial accounts will now be saved regardless. Lastly, a merit system was also implemented that notifies the user about new merits that they might have unlocked.