Implement Desync framework and create NPC graphics: v0.0.14
This update tweaks the NPC yet again, now integrating visuals for the first time. The way an NPC is displayed is based on the DialogueTone that is applied to the set Dialogue. Likewise, the Desync portion of Orikivo had more barebone classes implemented to help create a final structure for how the world progresses. An alternate game mode for the Casino, known as the Tick Machine has been added as its framework, and minor tweaks were made across several classes.