Apply event parsing and guild framework: v0.0.15
This minor update fixes the event parsing system to be dynamic and easier to apply. The new guild account was spruced up to now be correctly parsed in JSON, and minor tweaks were made to how merits were shown in the list, which now displays the rewards that you might receive from the merit. Lastly, minor tweaks were made across several classes.