Restructure Merits and Notifications: v0.1.5
This update focuses on adjusting the merit system previously used by WorldService, and places them in a more organized manner. MeritHandler now focuses on everything related to merits. The Notifier class was tweaked and made functional, now storing Notification values for a User, which can be displayed above a message, if applicable. Likewise, if the message is too large, the notifier holds off until it can be appended.