Add notes & begin final formatting framework
The goal for the final formatting framework is to make writing responses very simple. This will reduce the overall complexity needed to write out replies, as most of the base formatting applied to most methods seems to duplicate a lot of common format types. Page bars should also be automated.